requirements.txt
google-cloud-texttospeech==2.14.2
google-cloud-storage==2.11.0
requests==2.31.0